+++ src/clients/schemas.ts
@@ -17,7 +17,7 @@ const WorkflowSchema = z.object({
 });
 
 const JobSchema = z.object({
-  job_number: z.number(),
+  job_number: z.number().optional(),
   id: z.string(),
 });
 
@@ -39,6 +39,54 @@ const JobDetailsSchema = z.object({
     job_name: z.string(),
   }),
 });
+
+const FlakyTestSchema = z.object({
+  flaky_tests: z.array(
+    z.object({
+      job_number: z.number(),
+    }),
+  ),
+  total_flaky_tests: z.number(),
+});
+
+const TestSchema = z.object({
+  message: z.string(),
+  run_time: z.union([z.string(), z.number()]),
+  file: z.string().optional(),
+  result: z.string(),
+  name: z.string(),
+  classname: z.string(),
+});
+
+const PaginatedTestResponseSchema = z.object({
+  items: z.array(TestSchema),
+  next_page_token: z.string().nullable(),
+});
+
+const ConfigValidateSchema = z.object({
+  valid: z.boolean(),
+  errors: z
+    .array(
+      z.object({
+        message: z.string(),
+      }),
+    )
+    .nullable(),
+  'output-yaml': z.string(),
+  'source-yaml': z.string(),
+});
+
+export const Test = TestSchema;
+export type Test = z.infer<typeof TestSchema>;
+
+export const PaginatedTestResponse = PaginatedTestResponseSchema;
+export type PaginatedTestResponse = z.infer<typeof PaginatedTestResponseSchema>;
+
+export const FlakyTest = FlakyTestSchema;
+export type FlakyTest = z.infer<typeof FlakyTestSchema>;
+
+export const ConfigValidate = ConfigValidateSchema;
+export type ConfigValidate = z.infer<typeof ConfigValidateSchema>;
 
 // Export the schemas and inferred types with the same names as the original types
 export const Pipeline = PipelineSchema;