+++ src/tools/getFlakyTests/inputSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-export const getBuildFailureOutputInputSchema = z.object({
+export const getFlakyTestLogsInputSchema = z.object({
   workspaceRoot: z
     .string()
     .describe(
@@ -16,18 +16,11 @@ export const getBuildFailureOutputInputSchema = z.object({
         'For example: "https://github.com/user/my-project.git"',
     )
     .optional(),
-  branch: z
-    .string()
-    .describe(
-      'The name of the branch currently checked out in local workspace. ' +
-        'This should match local git branch. ' +
-        'For example: "feature/my-branch", "bugfix/123", "main", "master" etc.',
-    )
-    .optional(),
   projectURL: z
     .string()
     .describe(
       'The URL of the CircleCI project. Can be any of these formats:\n' +
+        '- Project URL: https://app.circleci.com/pipelines/gh/organization/project\n' +
         '- Project URL with branch: https://app.circleci.com/pipelines/gh/organization/project?branch=feature-branch\n' +
         '- Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123\n' +
         '- Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def\n' +